+++ erpnext/manufacturing/doctype/bom_update_tool/bom_update_tool.py
@@ -2,15 +2,14 @@
 # For license information, please see license.txt
 
 import json
-from typing import TYPE_CHECKING, Dict, Literal, Optional, Union
+from collections import defaultdict
+from typing import TYPE_CHECKING, Dict, List, Literal, Optional, Union
 
 if TYPE_CHECKING:
 	from erpnext.manufacturing.doctype.bom_update_log.bom_update_log import BOMUpdateLog
 
 import frappe
 from frappe.model.document import Document
-
-from erpnext.manufacturing.doctype.bom.bom import get_boms_in_bottom_up_order
 
 
 class BOMUpdateTool(Document):
@@ -47,7 +46,10 @@ def update_cost() -> None:
 	"""Updates Cost for all BOMs from bottom to top."""
 	bom_list = get_boms_in_bottom_up_order()
 	for bom in bom_list:
-		frappe.get_doc("BOM", bom).update_cost(update_parent=False, from_child_bom=True)
+		bom_doc = frappe.get_doc("BOM", bom)
+		bom_doc.calculate_cost(save_updates=True, update_hour_rate=True)
+		# bom_doc.update_exploded_items(save=True) #TODO: edit exploded items rate
+		bom_doc.db_update()
 
 
 def create_bom_update_log(
@@ -67,3 +69,90 @@ def create_bom_update_log(
 			"update_type": update_type,
 		}
 	).submit()
+
+
+def get_boms_in_bottom_up_order(bom_no: Optional[str] = None) -> List:
+	"""
+	Eg: Main BOM
+	        |- Sub BOM 1
+	                |- Leaf BOM 1
+	        |- Sub BOM 2
+	                |- Leaf BOM 2
+	Result: [Leaf BOM 1, Leaf BOM 2, Sub BOM 1, Sub BOM 2, Main BOM]
+	"""
+	leaf_boms = []
+	if bom_no:
+		leaf_boms.append(bom_no)
+	else:
+		leaf_boms = _get_leaf_boms()
+
+	child_parent_map = _generate_child_parent_map()
+	bom_list = leaf_boms.copy()
+
+	for leaf_bom in leaf_boms:
+		parent_list = _get_flat_parent_map(leaf_bom, child_parent_map)
+
+		if not parent_list:
+			continue
+
+		bom_list.extend(parent_list)
+		bom_list = list(dict.fromkeys(bom_list).keys())  # remove duplicates
+
+	return bom_list
+
+
+def _generate_child_parent_map():
+	bom = frappe.qb.DocType("BOM")
+	bom_item = frappe.qb.DocType("BOM Item")
+
+	bom_parents = (
+		frappe.qb.from_(bom_item)
+		.join(bom)
+		.on(bom_item.parent == bom.name)
+		.select(bom_item.bom_no, bom_item.parent)
+		.where(
+			(bom_item.bom_no.isnotnull())
+			& (bom_item.bom_no != "")
+			& (bom.docstatus == 1)
+			& (bom.is_active == 1)
+			& (bom_item.parenttype == "BOM")
+		)
+	).run(as_dict=True)
+
+	child_parent_map = defaultdict(list)
+	for bom in bom_parents:
+		child_parent_map[bom.bom_no].append(bom.parent)
+
+	return child_parent_map
+
+
+def _get_flat_parent_map(leaf, child_parent_map):
+	"Get ancestors at all levels of a leaf BOM."
+	parents_list = []
+
+	def _get_parents(node, parents_list):
+		"Returns recursively updated ancestors list."
+		first_parents = child_parent_map.get(node)  # immediate parents of node
+		if not first_parents:  # top most node
+			return parents_list
+
+		parents_list.extend(first_parents)
+		parents_list = list(dict.fromkeys(parents_list).keys())  # remove duplicates
+
+		for nth_node in first_parents:
+			# recursively find parents
+			parents_list = _get_parents(nth_node, parents_list)
+
+		return parents_list
+
+	parents_list = _get_parents(leaf, parents_list)
+	return parents_list
+
+
+def _get_leaf_boms():
+	return frappe.db.sql_list(
+		"""select name from `tabBOM` bom
+		where docstatus=1 and is_active=1
+			and not exists(select bom_no from `tabBOM Item`
+				where parent=bom.name and ifnull(bom_no, '')!='')"""
+	)
